fix: update test, rollback singleton
